backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv  # Optional, for environment variable management
